+++ src/index.html
@@ -15,8 +15,12 @@
       <load src="partials/home-page/hero.html" />
       <load src="partials/home-page/filters.html" />
       <load src="partials/home-page/home-content.html" />
+      <load src="partials/home-page/home-content/main-content.html" />
+      
+     
+
     </main>
       <load src="partials/components/footer.html" />
-    <script type="module" src="./js/main.js"></script>
+    <script type="module" src="./js/main.js" ></script>
   </body>
 </html>
