+++ src/js/main.js
@@ -2,6 +2,6 @@ import './home-content/home-content.js';
 import './filters/filters.js';
 import './modals/modal-product/modal-product.js';
 import './modals/modal-subscription/modal-subscription.js';
-import './cart-content/cart-content.js'
+
 
 console.log('5');
