--- src/partials/cart-page/cart-content/cart-products-list.html
@@ -1,5 +1,0 @@
-<section class="card-section">
-    <div class="container card-container">
-        
-    </div>
-</section>
