+++ src/partials/home-page/home-content/main-content.html
@@ -1,1 +1,8 @@
-<div class='main-content'></div>
+<div class='main-content'>
+    
+        <div class="container card-container">
+                
+            
+        </div>
+   
+</div>
